--- a/gpytorch/distributions/__init__.py
+++ b/gpytorch/distributions/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from .distribution import Distribution
+from .multivariate_normal import MultivariateNormal
+from .multitask_multivariate_normal import MultitaskMultivariateNormal
+
+
+__all__ = ["Distribution", "MultivariateNormal", "MultitaskMultivariateNormal"]