--- a/gpytorch/distributions/multivariate_normal.py
+++ b/gpytorch/distributions/multivariate_normal.py
@@ -114,6 +114,21 @@ class MultivariateNormal(TMultivariateNormal, Distribution):
         of the covariance matrix"""
         return self.loc + _batch_mv(self._unbroadcasted_scale_tril, base_samples)
 
+    def confidence_region(self):
+        """
+        Returns 2 standard deviations above and below the mean.
+
+        Returns:
+            Tuple[Tensor, Tensor]: pair of tensors of size (b x d) or (d), where
+                b is the batch size and d is the dimensionality of the random
+                variable. The first (second) Tensor is the lower (upper) end of
+                the confidence region.
+
+        """
+        std2 = self.stddev.mul_(2)
+        mean = self.mean
+        return mean.sub(std2), mean.add(std2)
+
     @property
     def variance(self):
         if self.islazy: