--- a/gpytorch/distributions/multivariate_normal.py
+++ b/gpytorch/distributions/multivariate_normal.py
@@ -41,30 +41,60 @@ class MultivariateNormal(TMultivariateNormal, Distribution):
                 loc=mean, covariance_matrix=covariance_matrix, validate_args=validate_args
             )
 
-    def __add__(self, other):
-        if isinstance(other, MultivariateNormal):
-            return self.__class__(
-                mean=self._mean + other.mean, covariance_matrix=self.covariance_matrix + other.covariance_matrix
-            )
-        elif isinstance(other, int) or isinstance(other, float):
-            return self.__class__(self.mean + other, self.covariance_matrix)
+    @property
+    def _unbroadcasted_scale_tril(self):
+        if self.islazy and self.__unbroadcasted_scale_tril is None:
+            # cache root decoposition
+            self.__unbroadcasted_scale_tril = self.covariance_matrix.root_decomposition()
+        return self.__unbroadcasted_scale_tril
+
+    @_unbroadcasted_scale_tril.setter
+    def _unbroadcasted_scale_tril(self, ust):
+        if self.islazy:
+            raise NotImplementedError("Cannot set _unbroadcasted_scale_tril for lazy MVN distributions")
         else:
-            raise RuntimeError("Unsupported type {} for addition w/ MultivariateNormal".format(type(other)))
+            self.__unbroadcasted_scale_tril = ust
 
-    def __radd__(self, other):
-        if other == 0:
-            return self
-        return self.__add__(other)
+    def confidence_region(self):
+        """
+        Returns 2 standard deviations above and below the mean.
 
-    def __truediv__(self, other):
-        return self.__mul__(1. / other)
+        Returns:
+            Tuple[Tensor, Tensor]: pair of tensors of size (b x d) or (d), where
+                b is the batch size and d is the dimensionality of the random
+                variable. The first (second) Tensor is the lower (upper) end of
+                the confidence region.
 
-    def __mul__(self, other):
-        if not (isinstance(other, int) or isinstance(other, float)):
-            raise RuntimeError("Can only multiply by scalars")
-        if other == 1:
-            return self
-        return self.__class__(mean=self.mean * other, covariance_matrix=self.covariance_matrix * (other ** 2))
+        """
+        std2 = self.stddev.mul_(2)
+        mean = self.mean
+        return mean.sub(std2), mean.add(std2)
+
+    def correlate_base_samples(self, base_samples):
+        """Correlate i.i.d. standard Normal samples using the root decomposition
+        of the covariance matrix"""
+        return self.loc + _batch_mv(self._unbroadcasted_scale_tril, base_samples)
+
+    @lazy_property
+    def covariance_matrix(self):
+        if self.islazy:
+            return self._covar
+        else:
+            return super(MultivariateNormal, self).covariance_matrix
+
+    def get_base_samples(self, sample_shape=torch.Size()):
+        """Get i.i.d. standard Normal samples (to be used with correlate_base_samples)"""
+        with torch.no_grad():
+            shape = self._extended_shape(sample_shape)
+            base_samples = _standard_normal(shape, dtype=self.loc.dtype, device=self.loc.device)
+        return base_samples
+
+    @property
+    def mean(self):
+        return self.loc
+
+    def representation(self):
+        return self.mean, self.covariance_matrix
 
     def rsample(self, sample_shape=torch.Size()):
         squeeze = True
@@ -85,50 +115,6 @@ class MultivariateNormal(TMultivariateNormal, Distribution):
 
         return res
 
-    def representation(self):
-        return self.mean, self.covariance_matrix
-
-    @property
-    def _unbroadcasted_scale_tril(self):
-        if self.islazy and self.__unbroadcasted_scale_tril is None:
-            # cache root decoposition
-            self.__unbroadcasted_scale_tril = self.covariance_matrix.root_decomposition()
-        return self.__unbroadcasted_scale_tril
-
-    @_unbroadcasted_scale_tril.setter
-    def _unbroadcasted_scale_tril(self, ust):
-        if self.islazy:
-            raise NotImplementedError("Cannot set _unbroadcasted_scale_tril for lazy MVN distributions")
-        else:
-            self.__unbroadcasted_scale_tril = ust
-
-    def get_base_samples(self, sample_shape=torch.Size()):
-        """Get i.i.d. standard Normal samples (to be used with correlate_base_samples)"""
-        with torch.no_grad():
-            shape = self._extended_shape(sample_shape)
-            base_samples = _standard_normal(shape, dtype=self.loc.dtype, device=self.loc.device)
-        return base_samples
-
-    def correlate_base_samples(self, base_samples):
-        """Correlate i.i.d. standard Normal samples using the root decomposition
-        of the covariance matrix"""
-        return self.loc + _batch_mv(self._unbroadcasted_scale_tril, base_samples)
-
-    def confidence_region(self):
-        """
-        Returns 2 standard deviations above and below the mean.
-
-        Returns:
-            Tuple[Tensor, Tensor]: pair of tensors of size (b x d) or (d), where
-                b is the batch size and d is the dimensionality of the random
-                variable. The first (second) Tensor is the lower (upper) end of
-                the confidence region.
-
-        """
-        std2 = self.stddev.mul_(2)
-        mean = self.mean
-        return mean.sub(std2), mean.add(std2)
-
     @property
     def variance(self):
         if self.islazy:
@@ -137,9 +123,27 @@ class MultivariateNormal(TMultivariateNormal, Distribution):
         else:
             return super(MultivariateNormal, self).variance
 
-    @lazy_property
-    def covariance_matrix(self):
-        if self.islazy:
-            return self._covar
+    def __add__(self, other):
+        if isinstance(other, MultivariateNormal):
+            return self.__class__(
+                mean=self._mean + other.mean, covariance_matrix=self.covariance_matrix + other.covariance_matrix
+            )
+        elif isinstance(other, int) or isinstance(other, float):
+            return self.__class__(self.mean + other, self.covariance_matrix)
         else:
-            return super(MultivariateNormal, self).covariance_matrix
+            raise RuntimeError("Unsupported type {} for addition w/ MultivariateNormal".format(type(other)))
+
+    def __radd__(self, other):
+        if other == 0:
+            return self
+        return self.__add__(other)
+
+    def __mul__(self, other):
+        if not (isinstance(other, int) or isinstance(other, float)):
+            raise RuntimeError("Can only multiply by scalars")
+        if other == 1:
+            return self
+        return self.__class__(mean=self.mean * other, covariance_matrix=self.covariance_matrix * (other ** 2))
+
+    def __truediv__(self, other):
+        return self.__mul__(1. / other)