This continues the work of #262 

- GPyTorch defines two custom distributions - `MultivariateNormal` and `MultitaskMultivariateNormal` - for use with LazyTensor covariance matrices.

Difference from original PR.
- Distributions define `rsample` for use with LazyTensors
- `rsample` and `sample` can take in base_samples - for correlated sampling
- `confidence_region` is now only a method on MVN and MTMVN
- `covariance_matrix` evaluates LazyTensors - `lazy_covariance_matrix` always returns a LazyTensor
- The `representation` method is gone for MVN (since it didn't do too much)
- Remove `RandomVariable`.

Closes #126