@gpleiss 2018-09-25
Fixed merge conflicts