@jacobrgardner 2018-09-25
Sorry if I'm just not caught up, but did we want to do anything to deprecate out at least `GaussianRandomVariable` since it's user facing? Something like:

```python
class GaussianRandomVariable(MultivariateNormal):
    def __init__(self, mean, covar): 
       warnings.warn("GaussianRandomVariable is deprecated in favor of ...", DeprecationWarning)
       super(GaussianRandomVariable, self).__init__(mean, covar)
    ...
    # Any other methods whose names changed call the right super method and throw a warning
```