@gpleiss 2018-09-25
Given how close beta is, I agree we can just add this breaking change to a release notes or something.