@gpleiss 2018-09-28
Timing results:

Standard GP: 0.643s -> 0.77s
AdditiveRegression test: 3.96s -> 3.95s
KroneckerRegression test: 10.3s -> 10.3s
CIFAR100 test: 0.617s (per minibatch)  -> 0.62s (per minibatch)