@Balandat 2018-09-28
Awesome, I expected it wouldn't change much especially w/ you changing this to calling `zero_mean_mvn_samples`, but better safe than sorry. Fire away!