@gpleiss 2018-09-28
Yeah it's definitely good to profile :) We should set up some actual profiling tests at some point